chore: 完成一系列兼容性修复与接口适配更新
1. 为CEGUI添加模态状态判断、屏蔽UI等兼容接口
2. 调整Win32编译选项添加/FS参数
3. 修复Cocos2d-x接口适配变更：
   - 替换废弃的CCScriptEngineManager等旧API
   - 修正HttpRequest类型枚举使用方式
   - 调整文件大小变量类型为ssize_t
   - 修正字符串编码转换相关问题
4. 完善MessageManager的多语言文本处理适配

+++ client/FireClient/Application/Manager/SpaceManager.h
@@ -5,6 +5,7 @@
 #include "network/HttpRequest.h"
 #include "network/HttpResponse.h"
 
+#include "platform/CCImage.h"
 #include "CEGUIImage.h" //dependencies\cegui\CEGUI\include\CEGUIImage.h
 #include <spine/Json.h>
 
@@ -26,7 +27,7 @@ public:
 	SpaceManager();
 	virtual ~SpaceManager();
 
-	void SendRequest(std::string strProtocolId, std::string strUrl, std::string strData = "", int nHttpType = cocos2d::network::HttpRequest::Type::POST, int nTimeOut = 5);
+	void SendRequest(std::string strProtocolId, std::string strUrl, std::string strData = "", int nHttpType = (int)cocos2d::network::HttpRequest::Type::POST, int nTimeOut = 5);
 	void ReceiveRequest_process(cocos2d::network::HttpClient* client, cocos2d::network::HttpResponse* response);
 
 	bool Initialize(const char* tempFilePath);
